+++ c/make_parens.py
@@ -2,25 +2,144 @@
 
 import sys
 
-def main():
-    """
-     this "doubles all parathensis"
-     to make new fingerprints
-     """
+class PermuteFingerprints(object):
+    def __init__(self):
+        self.fingerprints = set()
 
-    for line in sys.stdin:
-        fp = line.strip()
-        print fp
+    def aslist(self):
+        return sorted(list(self.fingerprints))
 
-        if fp.startswith('1'):
-            print 'v' + fp[1:]
-            print 's' + fp[1:]
-        if fp.startswith('s'):
-            print 'v' + fp[1:]
-            print '1' + fp[1:]
-        if fp.startswith('v'):
-            print 's' + fp[1:]
-            print '1' + fp[1:]
+    def insert(self,s):
+        if self.validate(s):
+            self.fingerprints.add(s)
+
+    def validate(self, s):
+        if len(s) == 0:
+            return False
+
+        if len(s) > 5:
+            s = s[0:5]
+
+        if 'X' in s:
+            return False
+
+        if '11' in s:
+            return False
+
+        if 'v1' in s:
+            return False
+
+        if 'nv' in s:
+            return False
+
+        # all 'ns' in union statements
+        if 'U' not in s and 'ns' in s:
+            return False
+
+        if 'nn' in s:
+            return False
+
+        if 'kk' in s:
+            return False
+
+        if 'ss' in s:
+            return False
+
+        if 'ooo' in s:
+            return False
+
+        if 'vvv' in s:
+            return False
+
+        # folded away
+        if '1o1' in s:
+            return False
+        if '1on' in s:
+            return False
+        if 'no1' in s:
+            return False
+        if 'non' in s:
+            return False
+        if '1(v' in s:
+            return False
+        if '1(n' in s:
+            return False
+        if '1(s' in s:
+            return False
+        if '1(1' in s:
+            return False
+        if 's(s' in s:
+            return False
+        if 's(n' in s:
+            return False
+        if 's(1' in s:
+            return False
+        if 's(v' in s:
+            return False
+        if 'v(s' in s:
+            return False
+        if 'v(n' in s:
+            return False
+        if 'v(1' in s:
+            return False
+        if 'v(v' in s:
+            return False
+
+        if s.startswith('vs'):
+            return False
+
+        if ')(' in s:
+            return False
+
+        # need to investigate E(vv) to see
+        # if it's correct
+        if 'vv' in s and s != 'E(vv)':
+            return False
+
+        # bogus
+        if s in ('sns', '1&n', 's1s', '1n1', '1o1', '1os'):
+            return False
+
+        # unlikely to be sqli but case FP
+        if s in ('so1n)', 'sonoE'):
+            return False
+
+        return True
+
+    def permute(self, fp):
+
+        self.insert(fp)
+
+        # do this for safety
+        if len(fp) > 1 and len(fp) < 5 and fp[-1] != 'c':
+            self.insert(fp + "c")
+
+        for i in range(len(fp)):
+            if fp[i] == '1':
+                self.insert(fp[0:i] + 'v' + fp[i+1:])
+                self.insert(fp[0:i] + 's' + fp[i+1:])
+                self.insert(fp[0:i] + 'f(1)' + fp[i+1:])
+                self.insert(fp[0:i] + '1os' + fp[1:])
+                self.insert(fp[0:i] + '1ov' + fp[1:])
+                self.insert(fp[0:i] + '1on' + fp[1:])
+            elif fp[i] == 's':
+                self.insert(fp[0:i] + 'v' + fp[i+1:])
+                self.insert(fp[0:i] + '1' + fp[i+1:])
+                self.insert(fp[0:i] + 'f(1)' + fp[i+1:])
+                self.insert(fp[0:i] + 'so1' + fp[1:])
+                self.insert(fp[0:i] + 'sov' + fp[1:])
+                self.insert(fp[0:i] + 'son' + fp[1:])
+            elif fp[i] == 'v':
+                self.insert(fp[0:i] + 's' + fp[i+1:])
+                self.insert(fp[0:i] + '1' + fp[i+1:])
+                self.insert(fp[0:i] + 'f(1)' + fp[i+1:])
+                self.insert(fp[0:i] + 'vo1' + fp[1:])
+                self.insert(fp[0:i] + 'vos' + fp[1:])
+                self.insert(fp[0:i] + 'von' + fp[1:])
+            elif fp[i] == ')':
+                self.insert(fp[0:i] + '))' + fp[1:])
+                self.insert(fp[0:i] + ')))' + fp[1:])
+                self.insert(fp[0:i] + '))))' + fp[1:])
 
         if '(' in fp:
 
@@ -31,8 +150,8 @@ def main():
                     parts.append(c)
                     done = True
                 parts.append(c)
-            newline = ''.join(parts)[0:5]
-            print newline
+            newline = ''.join(parts)
+            self.insert(newline)
 
             done = False
             parts = []
@@ -43,8 +162,8 @@ def main():
                     else:
                         done = True
                 parts.append(c)
-            newline = ''.join(parts)[0:5]
-            print newline
+            newline = ''.join(parts)
+            self.insert(newline)
 
             done = False
             parts = []
@@ -52,11 +171,22 @@ def main():
                 if c == '(':
                     parts.append(c)
                 parts.append(c)
-            newline = ''.join(parts)[0:5]
-            print newline
+            newline = ''.join(parts)
+            self.insert(newline)
 
 
 if __name__ == '__main__':
-    main()
+    """
+     this "doubles all parathensis"
+     to make new fingerprints
+     """
+
+    mutator = PermuteFingerprints()
+
+    for line in sys.stdin:
+        mutator.insert(line.strip())
+
+    for fp in mutator.aslist():
+        print fp